+++ efs-io/src/main/java/org/efs/io/EfsRow.java
@@ -16,7 +16,10 @@
 
 package org.efs.io;
 
+import jakarta.annotation.Nonnull;
 import java.time.Instant;
+import java.util.Set;
+import javax.annotation.concurrent.Immutable;
 import org.efs.event.IEfsEvent;
 
 /**
@@ -28,6 +31,7 @@ import org.efs.event.IEfsEvent;
  * @author <a href="mailto:rapp@acm.org">Charles W. Rapp</a>
  */
 
+@Immutable
 public final class EfsRow<E extends IEfsEvent>
     implements IEfsEvent
 {
@@ -50,6 +54,13 @@ public final class EfsRow<E extends IEfsEvent>
     private final long mRowIndex;
 
     /**
+     * Immutable set containing event row tags. These tags are
+     * user-defined. This set may be empty but may not contain
+     * {@code null} values.
+     */
+    private final Set<Integer> mTags;
+
+    /**
      * Published event itself.
      */
     private final E mEvent;
@@ -67,16 +78,19 @@ public final class EfsRow<E extends IEfsEvent>
      * timestamp, row index, and event.
      * @param timestamp publish timestamp.
      * @param rowIndex row index.
+     * @param tags user-defined tags associated with this row.
      * @param event actual efs event.
      */
     /* package */ EfsRow(final Instant timestamp,
                          final long rowIndex,
+                         final Set<Integer> tags,
                          final E event)
     {
         mPublishTimestamp = timestamp;
         mRowIndex = rowIndex;
+        mTags = tags;
         mEvent = event;
-    } // end of EfsRow(Instant, int, E)
+    } // end of EfsRow(Instant, int, Set<>, E)
 
     //
     // end of Constructors.
@@ -113,6 +127,7 @@ public final class EfsRow<E extends IEfsEvent>
      * {@link EfsFile}.
      * @return event publish timestamp.
      */
+    @Nonnull
     public final Instant getPublishTimestamp()
     {
         return (mPublishTimestamp);
@@ -128,9 +143,21 @@ public final class EfsRow<E extends IEfsEvent>
     } // end of getRowIndex()
 
     /**
+     * Returns user-define tags immutable set. Set may be empty
+     * but does not contain {@code null} values.
+     * @return immutable set of user defined tags.
+     */
+    @Nonnull
+    public final Set<Integer> getTags()
+    {
+        return (mTags);
+    } // end of getTags()
+
+    /**
      * Returns event stored in {@link EfsFile}.
      * @return stored event.
      */
+    @Nonnull
     public final E getEvent()
     {
         return (mEvent);